mir_ci: reorganize project
Place things into a more sane structure:

├── mir_ci
│   ├── fixtures
│   ├── interfaces
│   ├── lib
│   ├── program
│   ├── tests
│   │   ├── clients
│   │   ├── data
│   │   └── robot
│   │       ├── libraries
│   │       ├── resources
│   │       └── templates
│   └── wayland
│       └── data
└── tools
