+++ mir-ci/hatch_build.py
@@ -18,7 +18,7 @@ def generate_protocol(path: Path, name: str, imports: Dict[str, str]) -> Dict[st
 
 class CustomBuildHook(BuildHookInterface):
     def initialize(self, version, build_data):
-        path = Path(__file__).parent / "mir_ci"
+        path = Path(__file__).parent / "mir_ci" / "wayland"
         core_imports = generate_protocol(path, "wayland", {})
         generate_protocol(path, "wlr-screencopy-unstable-v1", core_imports)
         generate_protocol(path, "wlr-virtual-pointer-unstable-v1", core_imports)
