+++ mir-ci/mir_ci/interfaces/benchmarkable.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from mir_ci.cgroups import Cgroup
+from ..lib.cgroups import Cgroup
 
 
 class Benchmarkable(ABC):
