--- mir-ci/mir_ci/protocols/__init__.py
@@ -1,4 +1,0 @@
-from .wayland import *
-from .wlr_screencopy_unstable_v1 import *
-from .wlr_virtual_pointer_unstable_v1 import *
-from .xdg_output_unstable_v1 import *
