+++ mir-ci/mir_ci/tests/clients/drag_and_drop_demo.py
