+++ mir-ci/mir_ci/tests/clients/maximizing_gtk_app.py
