+++ mir-ci/mir_ci/tests/robot/templates/drag_and_drop_end.png
