+++ mir-ci/mir_ci/wayland/data/wlr-virtual-pointer-unstable-v1.xml
