+++ mir-ci/mir_ci/wayland/protocols/__init__.py
@@ -1,0 +1,4 @@
+from .wayland import *  # noqa: F401,F403
+from .wlr_screencopy_unstable_v1 import *  # noqa: F401,F403
+from .wlr_virtual_pointer_unstable_v1 import *  # noqa: F401,F403
+from .xdg_output_unstable_v1 import *  # noqa: F401,F403
