+++ mir-ci/mir_ci/wayland/screencopy_tracker.py
@@ -5,13 +5,13 @@ import os
 import stat
 from typing import Any, Dict, Optional
 
-from mir_ci.protocols import WlOutput, WlShm, ZwlrScreencopyManagerV1
-from mir_ci.protocols.wayland.wl_buffer import WlBufferProxy
-from mir_ci.protocols.wayland.wl_output import WlOutputProxy
-from mir_ci.protocols.wayland.wl_shm import WlShmProxy
-from mir_ci.protocols.wlr_screencopy_unstable_v1.zwlr_screencopy_frame_v1 import ZwlrScreencopyFrameV1Proxy
-from mir_ci.protocols.wlr_screencopy_unstable_v1.zwlr_screencopy_manager_v1 import ZwlrScreencopyManagerV1Proxy
-from mir_ci.wayland_client import WaylandClient
+from .protocols import WlOutput, WlShm, ZwlrScreencopyManagerV1
+from .protocols.wayland.wl_buffer import WlBufferProxy
+from .protocols.wayland.wl_output import WlOutputProxy
+from .protocols.wayland.wl_shm import WlShmProxy
+from .protocols.wlr_screencopy_unstable_v1.zwlr_screencopy_frame_v1 import ZwlrScreencopyFrameV1Proxy
+from .protocols.wlr_screencopy_unstable_v1.zwlr_screencopy_manager_v1 import ZwlrScreencopyManagerV1Proxy
+from .wayland_client import WaylandClient
 
 libc = ctypes.cdll.LoadLibrary(None)  # type: ignore
 shm_counter = 0
