+++ mir-ci/tools/virtual_pointer_debug.py
